% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainyR-package.R
\docType{package}
\name{brainyR-package}
\alias{brainyR}
\alias{brainyR-package}
\title{brainyR: Gene expression information for purified brain signatures}
\description{
The brainyR package contains dataframe of gene expression information of purified brain cell-types as collated and benchmarked in Sutton et al. (2021)
}
\keyword{internal}
