% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainyR-package.R
\docType{package}
\name{brainyR-package}
\alias{brainyR}
\alias{brainyR-package}
\title{brainyR: Gene Expression Information for Purified Brain Signatures}
\description{
Dataframe of gene expression information of purified brain cell-types.
}
\keyword{internal}
