% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSignature.R
\name{loadSignature}
\alias{loadSignature}
\title{Load a specific brain dataset}
\usage{
loadSignature(dataset, cell_types)
}
\arguments{
\item{dataset}{dataset of interest. Options include: F5, IP, SC, MM, LK, CA}

\item{cell_types}{cell types of interest. Options include: Neurons, Astrocytes, Oligodendrocytes, Microglia, Endothelia}
}
\value{
Object with options from dataset and cell_type
}
\description{
This function loads a brain signature from given datasets and for the given cell type.
It goes through the sigsBrain.rda file and returns the relevant signatures.
If signature is not in the given .rda file, it will return an error message.
}
\examples{
loadSignature("F5", "Neurons")
loadSignature("F5", "all")
}
